extracted keypad.h
